requirements.txt
tqdm==4.60.0
networkx==2.5.1
torch==1.7.1+cu101
matplotlib==3.3.4
scikit_image==0.17.2
torchvision==0.8.2+cu101
numpy==1.19.5
torch_geometric==1.7.0
dotmap==1.3.23
pandas==1.1.5
gensim==4.0.1
plotly==4.9.0
Pillow==8.2.0
scikit_learn==0.24.2
skimage==0.0
